requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.1.0
base58==2.1.1
bip-utils==2.8.0
bip32==3.4
bip44==0.1.4
bitarray==2.8.3
cbor2==5.5.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
coincurve==18.0.0
crcmod==1.7
cytoolz==0.12.2
ecdsa==0.18.0
ed25519-blake2b==1.4
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.2
eth-utils==2.3.1
frozenlist==1.4.0
hexbytes==0.3.1
idna==3.4
importlib-resources==6.1.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
lru-dict==1.2.0
mnemonic==0.20
multidict==6.0.4
parsimonious==0.9.0
pkgutil-resolve-name==1.3.10
protobuf==4.25.1
py-sr25519-bindings==0.2.0
pycparser==2.21
pycryptodome==3.19.0
PyNaCl==1.5.0
python-dotenv==1.0.0
pyunormalize==15.1.0
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
rlp==3.0.0
rpds-py==0.13.1
six==1.16.0
toolz==0.12.0
typing-extensions==4.8.0
urllib3==2.1.0
web3==6.11.3
websockets==12.0
yarl==1.9.3
zipp==3.17.0
